+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [v0.9.0]
 ## Features
 - error basic struct used for typical error return value
-- background goroutine loop task controller with channel sign and sync.WaitGroup
+- [status package] background goroutine loop task controller with channel sign and sync.WaitGroup
+- [element package] item container that support simple database features.